+++ R/Distribution.R
@@ -826,8 +826,8 @@ Distribution$set("public","cdf",function(x1, ..., lower.tail = TRUE, log.p = FAL
     cdf = numeric(length(x1))
     cdf[x1 > self$sup()] = 1
 
-    if(getR6Class(self) == "Empirical"){
-      if(any(x1 >= self$inf()) | any(x1 <= self$sup()))
+    if(getR6Class(self) %in% c("Empirical","WeightedDiscrete")){
+      if(any(x1 >= self$inf() & x1 <= self$sup()))
         cdf[x1 >= self$inf() | x1 <= self$sup()] = private$.cdf(x1[x1 >= self$inf() | x1 <= self$sup()])
     } else {
       if(any(self$liesInSupport(x1, all = F)))